Add a license
It looks like there's no license on this project, I would assume this is because you meant for it to be open source, however if nothing is specified it is by default closed-source software.  I would suggest an MIT, or perhaps LGPL license for this project.  If you want more information check out